lab_07/block_products_app/requirements.txt
faust-streaming[rocksdb]
confluent-kafka
python-schema-registry-client
cachetools
authlib
